pyproject.toml
[project]
name = "adk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.12.0",
    "psycopg2>=2.9.10",
    "pydantic>=2.11.7",
    "yfinance>=0.2.65",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["tools*"]
